done with validation

+++ src/main/java/hr/algebra/tabletopshop/dto/BoardgameDto.java
@@ -3,5 +3,5 @@ package hr.algebra.tabletopshop.dto;
 import java.math.BigDecimal;
 
 public record BoardgameDto(Integer Id, String name, Integer year, Double rating, Double weight, Integer playingTime,
-                           BigDecimal price) {
+                           Double price) {
 }
